moved the getNextRecurrence date to it's own file and imported it separately into the recurrence.helper file and the meeting scheduler. Also changed the currentDate function. All seems to work now
